hw7/requirements.txt
wheel
numpy==1.21.4
gym==0.17.3
free-mujoco-py==2.1.6
pybullet==3.2.0
tensorboard==2.3.0
tensorboardX==1.8
matplotlib
ipython==6.4.0
#moviepy==1.0.0
pyvirtualdisplay==1.3.2
torch
torchvision
opencv-python
ipdb==0.13.3
hydra-core==1.1.1
# protobuf==3.20
networkx # ==2.5
pytest==7.4.3
pandas==2.0.3
chardet==5.2.0
comet_ml

gin-config
setuptools
wheel
gin-config
stable-baselines3
gymnasium
protobuf==3.20.*
decorator==4.0.2
moviepy
git+https://github.com/montrealrobotics/puppersim.git@charlie_workdir
